Implement partially broken deserialization.
This may or may not work with the `CQRSSerializer(data=…)` technique; I
haven't tried it at all.

Unfortunately it appears that somewhere, deep in the bowels of Django
REST framework, probably, the `mongoID` field is lost in
deserialization. Hence the test_c_deserialize and test_d_deserialize
tests fail. To me this is inexplicable; I've tried somewhat to
understand it or to work around it, but in vain.