Rewrite the serializer tests completely.
The structure that I had was this:

    CQRSPolymorphicModel
       |     |
       A     B
       |\_ _/|
       \  C  /
        \_ _/
          D

… with A and D having manually specified serializers and B and C
automatic.

This let me feel clever about having taken care of the edge cases of
diamond inheritance, but it's actually not a sound thing; Django *seems*
to allow it, but it's certainly going to have problems and you shouldn't
ever write it.

So I threw it all away and replaced it with this structure:

             CQRSPolymorphicModel
             /                 \
            A                   M
         __/ \___            __/ \___
       AA        AM        MA        MM
      /  \      /  \      /  \      /  \
    AAA  AAM  AMA  AMM  MAA  MAM  MMA  MMM

(Three levels is probably enough, but a fourth would be entirely
possible.)

The left hand (A) branches have automatic serializers and the right hand
(M) branches have manual serializers.

Everything now passes. The mongoID being lost on deserialisation was
indeed an artefact of the diamond inheritance.