At last! Throw away our custom Mongo stuff.
That is,

(a) the mongo backend, and
(b) the CQRSModelMixin with its mongoID field.

The work of the past couple of weeks has finally finished making it all
obsolete.

Now is ushered in a new reign of peace and serenity, untroubled by
Mongo-specific hacks. (Just all the rest of my hacks :P.)